requirements.txt
feedgen
pyyaml
easydict